installer: when looking for running Git processes, look for DLLs
Instead of looking for, say, `git.exe`, we can look for processes using
`zlib1.dll`. That will catch `git.exe` processes and in addition other
Git processes such as `git-remote-https.exe` (which we failed to catch
previously).

While at it, remove the explicit `.exe` files for which we looked; These
are all covered by the `.dll` files we are already looking for. This is
now possible because we use the Restart Manager API exclusively now
(earlier, we could not, because we still supported pre-Vista Windows
versions, but we dropped XP support a long time ago).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>