installer: try to stop running FSMonitor daemons
If there are running Git processes that would block upgrading Git for
Windows, they _might_ be FSMonitor daemons.

Let's deal with this situation by enumerating the named pipes opened by
FSMonitor daemons, deducing the worktree paths from the named pipes'
names, and trying to stop those daemons.

Technically, this might stop too many FSMonitor daemons. After all, it
is possible to have multiple Git for Windows versions side by side e.g.
via installing a Portable Git or a Git for Windows SDK.

However, there are no trivial ways to avoid stopping FSMonitor daemons
that do not actually use the files that are part of the Git for Windows
installation we are about to upgrade.

The most obvious idea would be to determine the full command-line of the
blocking Git processes, then ignore those that are not FSMonitor
daemons, then determine those processes' respective current working
directories, and then issuing `git fsmonitor--daemon stop` in those
working directories.

However, there is no Win32 API for determining the command-line
for a process given by its ID (see e.g.
https://devblogs.microsoft.com/oldnewthing/20091125-00/?p=15923). We
_could_ call `OpenProcess()` and then `ZwReadVirtualMemory()` to read
the Process Environment Block (where the command-line is stored), but
anti-malware software might mistake the installer for some malware if it
does that.

Likewise, obtaining the current working directory seems to be possible
only via injecting a remote thread into the running process, calling
`GetCurrentDirectoryW()`, which anti-malware would be prone to mistake
for something we shouldn't do.

Yet another approach would have been to abuse the Restart Manager API by
modifying the `fsmonitor--daemon` command to register itself as
restartable, with a command-line that could be queried via the Restart
Manager API, with the ultimate goal of getting the working directory.
But that would not be backwards-compatible, and it might present a
challenge trying to get that accepted in upstream Git, anyway, as it is
not really meant to be restartable.

So let's punt on all this and simply shut down _all_ FSMonitor daemons
if Git for Windows' installer detects that any of Git for Windows' files
are in use.

Note: while the current FSMonitor uses a pipe name that ends in
`\fsmonitor--daemon.ipc`, the first iteration used a name that ends in
`\fsmonitor`, therefore we have to be a bit careful to catch both.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>