installer: support shutting down older FSMonitor daemons
Git for Windows shipped with v1 of that experimental feature, where the
subcommands were still offered as command modes. The difference is that
older FSMonitor daemons (such as the ones shipped in Git for Windows
v2.31.0) need to be shut down using the `--stop` option (with two
leading dashes) while newer ones need the `stop` command (without any
dashes).

Let's obtain the usage text and deduce which form to use dynamically.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>