installer: try to shut down each FSMonitor only once
When we failed to shut down an FSMonitor in a given worktree, or when
the user works in that worktree while upgrading Git and thereby
inadvertently restarts the daemon, let's not try to shut it down again.
Let the user handle it manually instead (and do not nag them about it).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>