installer: warn about switching away from .vfs versions
The .vfs versions of Git for Windows are aware not only of VFS for Git,
but they also come with the `gvfs-helper` command that can be used to
emulate partial clones with Azure Repos.

This means that users switching away from .vfs versions might end up
with broken VFS for Git or Scalar enlistments.

Let's give them a chance to avoid that.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>