please.sh upgrade: learn about pkgconf
The MSYS2 project switched from the original `pkg-config` package to the
`pkgconf` package (which supersedes the former) a long time ago.

Unfortunately, at that point they already had stopped servicing the i686
side of MSYS2 (except for MINGW packages).

But we do need a newer `pkgconf` version to be able to build a newer
`libgpg-devel` (which relies on `pkg-config <package> --version` to
report the version of the package, not of `pkg-config`, as our current
i686 version of `pkg-config.exe` does).

So let's teach the `please.sh` script how to build `pkgconf`, then
switch Git for Windows' i686 SDK over.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>