Enable tab-completion of `git update-git-for-windows`
Things that are tucked away in `libexec/git-core/` and _not_ listed as
main porcelain in Git's `command-list.txt` file won't show up when
hitting the `Tab` key.

However, the `git update-git-for-windows` command is _really_ useful and
should be more discoverable.

So let's just install the script into the `bin/` directory instead,
which is the proper place for add-on Git commands, anyway.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>