make-file-list: exclude bogus libsqlite dependency
In https://github.com/msys2/MSYS2-packages/commit/325eedfc14 (a rather
monster-type of a huge change), the build-time dependency `tcl` was
upgraded to a full-scale runtime dependency of all SQLite packages,
including `libsqlite`.

This is most likely a bug (which was easy to miss, given the amount of
changes accumulated in that big patch), as `tcl` is still marked as a
runtime dependency of the `tcl-sqlite` package (which is still correct).

But `tcl` is not required to use `libsqlite`, which `openssh` does, and
therefore we now have this useless baggage we have to shlep around.

Let's exclude it manually.

Noticed in
https://github.com/git-for-windows/git/issues/1439#issuecomment-867310152

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>