MinGit: use the Git wrapper as ash.exe, if possible
Some programs, such as OpenSSH, require the environment variable `SHELL`
to contain an absolute path to an executable. No extra arguments
allowed.

However, BusyBox' `ash` would need to be called like this:

	<mingit>\mingw64\bin\busybox.exe sh [...]

If the `sh` is omitted, only mayhem will ensue.

To allow for setting `SHELL` in such a manner, we taught the good ole'
Git wrapper to stand in. That is, if the Git wrapper detects that it is
run with the base name `ash.exe` (or `sh.exe`), it will try to spawn
BusyBox' `ash` with the same argument, wait for the command to conclude,
and then exit with the same exit code.

Let's use this version of the Git wrapper, if we have it.

This requires https://github.com/git-for-windows/MINGW-packages/pull/52
to work, and allows e.g. using a `ProxyCommand` with OpenSSH, after
setting `SHELL` to point to the `ash.exe` file.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>