installer: be more verbose in the post-install step
It can be a bit unnerving when installing Git for Windows and the
progress bar hit the end, and nothing seems to happen for a while.

A lot is happening.

Just like we did in the recent feature where we report what we are doing
during a Scalar upgrade from Scalar.NET to Scalar/C, let's report what
we are doing during the remainder of the `ssPostInstall` phase.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>