installer: avoid using `cmd.exe` just to capture stdout/stderr
We just introduced a new helper function that allows capturing the
output into a `String` directly, without having to run it through
`cmd.exe`.

We need to do this without `cmd.exe` because there have been reports
that it does not work quite as designed. See e.g.
https://github.com/git-for-windows/git/issues/3226 where it looks as if
in certain locked-down setups, running `cmd.exe` will not be allowed.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>