installer: Specify name of app in Control Panel
The version of the app is duplicated in the name of the app in the list of installed apps in "Settings - Apps - Apps & features" and "Control Panel - Programs - Programs and Features" in the version for Windows.

To remove the version number from the application name, the UninstallDisplayName value has been added to the [Setup] section, as recommended in the documentation https://jrsoftware.org/ishelp/index.php?topic=setup_uninstalldisplayname.

Closes https://github.com/git-for-windows/git/issues/3319.

Signed-off-by: Pavel Moiseenko <rakleed@gmail.com>