installer: stay responsive while uninstalling the previous Git
The uninstaller can take a while to run, and when called using the
current way, the installer window will become unresponsive (for
example, it cannot be moved by the user).

To counter that, use the `ExecWithCapture()` function, which lets the
UI stay responsive, and as a bonus, show the `stderr` in case of a
failure.

While at it, indicate what we are doing while uninstalling the previous
Git for Windows version.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>