installer: still only offer PuTTY if at least one session was found
Previously, we only showed the PuTTY page if at least one session was
found. But in the meantime, we renamed the page and offer to use any
`ssh.exe` found on the `PATH` as an alternative.

Which means that we now also offer PuTTY if such an `ssh.exe` was found,
even if there are no saved PuTTY sessions.

Let's hide the PuTTY option unless at least one PuTTY session was found.

This patch has to re-indent a lot of lines, therefore the diff is best
viewed with the `-w` option.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>