installer: shut down FSMonitors in the worktrees, not gitdirs
The FSMonitor-pre-v4 version we integrated into v2.33.0-rc1 wants to
ensure that it runs only in worktrees, not in bare repositories.

Unfortunately, when starting `git.exe fsmonitor--daemon` in a `.git`
directory, it will claim that it is a bare repository.

Even more unfortunate, this also happens when we call `git.exe
fsmonitor--daemon stop` in a gitdir which we _know_ to have the named
pipe, i.e. we know that an FSMonitor daemon is running there.

Let's work around this by ensuring that we always call this command in
the worktree associated with the gitdir, not in the gitdir itself.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>