installer: handle IsDowngrade()/IsUpgrade() with hard-coded versions
To determine whether to show the components page under "only show unseen
options", or whether to auto-upgrade Git Credential Manager for Windows
users to Git Credential Manager Core, we call `IsUpgrade(<version>)`
with a hard-coded version, to find out whether the still-installed
version is older than the specified version.

However, the fine-grained logic that tried to dig deeper upon -rc
versions made that check incorrect: in case the still-installed version
is an -rc version, we replace the hard-coded version by the output of
`git version` at the time the installer was built (which is pretty much
definitely incorrect at this point).

Let's be more careful to look at that snapshot of the output of `git
version` _only_ if `IsDowngrade()` or `IsUpgrade()` was called with the
version that is about to be installed.

If not, let's fake some output. For example, when looking whether the
currently-installed version of Git for Windows would upgrade to v2.29.0
via `IsUpgrade('2.29.0')`, we replace the hard-coded string by `git
version 2.29.0.windows.1`.

This fixes the problem where the v2.33.0-rc* installers _still_ showed
the components page (because of the Windows Terminal Profile option,
which was introduced into v2.32.0), even if the user asked to be shown
only unseen options.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>