installer: do show the "external SSH" choice if unseen
The "external SSH" choice was introduced in v2.33.0, and piggy-backs on
the previous version of the SSHChoice page that only showed Tortoise
Plink as an option (and only if sessions were stored already).

Therefore, this page is not marked as "unseen", but we do want to show
it to users. Let's do that if they are upgrading from any version prior
to v2.33.0.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>