installer: avoid <() Bashism
When building installers interactively, this works because Git for
Windows' SDK's default shell is Bash.

However, in the automated build of Git for Windows, `sh.exe` is used,
which is technically identical to `bash.exe` but apparently behaves
differently:

	$ sh -c 'comm -12 <(printf "%s\n" 1 2) <(printf "%s\n" 2 3)'
	sh: -c: line 0: syntax error near unexpected token `('
	sh: -c: line 0: `comm -12 <(printf "%s\n" 1 2) <(printf "%s\n" 2 3)'

	$ bash -c 'comm -12 <(printf "%s\n" 1 2) <(printf "%s\n" 2 3)'
	2

As a consequence, we saw this problem in the build of v2.33.0-rc2 (see
https://dev.azure.com/git-for-windows/git/_build/results?buildId=82662&view=logs&j=fd84893c-d8cf-515f-6f6b-748e5f0b9fc1&t=53291ad3-d296-5424-da40-80462f49966c&l=31232):

	.../release.sh: command substitution: line 284: syntax error near unexpected token `('
	.../release.sh: command substitution: line 284: `comm -12 <(echo "$LIST" | sort) <(pacman -Ql openssh | sed -n 's|^openssh /\(.*[^/]\)$|\1|p' | sort) |'

Let's just not use the `<()` construct at all.

Reported by Carlo Marcelo Arenas Belón.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>