git-extra: include an simple Win32 `git askpass`
This is a drop-in replacement for `git gui--askpass`. Since we added an
option to use an external `ssh` found on the `PATH` (#367) we'll need an
`askpass` implementation that can be called by any windows application even
without understanding shebang lines. `git gui--askpass` probably also has
the same problems with screenreaders that `git gui--askyesno` had(#234), so
we'll likely get improved accessibility as a positive side-effect.

Signed-off-by: Matthias Aßhauer <mha1993@live.de>