installer: fix the code that wanted to mark ssh-agent as "to terminate"
When Git for Windows' files are in use, they cannot be replaced. For
that reason, Git for Windows' installer lists the processes that use
those files, and in some instances offers to automatically terminate
them (e.g. ssh-agent, because it is deemed okay to just terminate it, it
can be auto-started again after the upgrade).

However, this never worked correctly with the code that uses the Restart
Manager API. Let's fix this.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>