installer: fix the question when asking to terminate some processes
To be able to replace files during an upgrade, we might ask the user
whether they are ready to go ahead with stopping/restarting certain
applications that are deemed relatively safe to stop automatically.

However, the question we asked assumed that we would be able to restart
the application, which is only possible for those processes that are
registered with the Restart Manager API. But processes such as the GPG
agent are not, they are only terminated, but not restarted.

Phrase the question differently, so that we stop giving the false
impression that we will restart also those applications that we cannot
restart.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>