sdk: warn when MSYS_NO_PATHCONV is set
When the environment variable `MSYS_NO_PATHCONV` is set, the MSYS2
runtime will stop trying to convert environment variables and
command-line arguments it detects as Unix-style paths to Windows paths
when executing non-MSYS2 programs.

This can be very helpful e.g. when calling `git blame -L /<some-regex>/
...` so that the regular expression is not mistaken for a Unix-style
absolute path and transformed.

However, setting that environment variable globally is prone to wreak
havoc, and we should warn loudly when detecting this situation.

This addresses the concern identified as the root cause of
https://github.com/git-for-windows/git/issues/3381

Signed-off-by: Alex <greenozon@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>