vimrc: avoid unexpected behavior if user remapped g`
In the autocmd

    autocmd BufReadPost *
      \ if line("'\"") > 0 && line("'\"") <= line("$")
      \           && expand("%") !~ "COMMIT_EDITMSG"
      \           && expand("%") !~ "ADD_EDIT.patch"
      \           && expand("%") !~ "addp-hunk-edit.diff"
      \           && expand("%") !~ "git-rebase-todo" |
      \   exe "normal g`\"" |
      \ endif

the command normal g` causes errors if the user
rebound in Vim the rarely used key sequence g` by :n(nore)nap g` ...
to a custom command.

This is fixed  by replacing normal g` with normal! g` which ignores any
user defined remappings and thus always achieves what is intended,
jumping to the last known position in a file (see :help g` in Vim).

Users of Vim on the Linux distribution openSUSE face the same issue:

  https://bugzilla.opensuse.org/show_bug.cgi?id=1190332

Signed-off-by: Enno Nagel enno.nagel+git@gmail.com
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>