git-update-git-for-windows: strip 'v' from fork tag
When checking for a Git version from a fork such as microsoft/git, we
need to extract the tag name from a JSON object that we get from a REST
API. This differs from the git-for-windows/git mechanism of using the
gitforwindows.org/latest-tag.txt file. In the case of the latest-tag.txt
file, the letter 'v' is stripped from the tag name, allowing comparison
to the installed Git version, which does not include the 'v'.

The 'v' was not previously stripped from the tag name from the fork.
Rectify that, allowing comparisons that present an upgrade.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>