pacman-helper.sh quick_add: allow uploading `-any` packages
We are about to build and upload the `ca-certificates` package for i686
(because the MSYS2 project no longer builds i686 versions of the MSYS
packages). This package is not marked with an architecture, though,
therefore we need to have a way to specify that we want to deploy it to
the i686 part of our Pacman repository.

With this change, we allow that, by specifying the architecture via the
environment variable `FALLBACK_ARCHITECTURE`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>