installer: add a script to help run down the pre-flight check list
It is not complete, as part of the check list are validating GUI things
(such as: `git gui` opens the Git GUI window and its contents look
sensible). But at least a couple of things are automated, leaving less
room for human error or sloppiness (such as myself missing something
after a full day fighting with an unexpected upstream version and cloud
outages).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>