added different return code for yes on update
If the user selects yes the update code is not waiting for the update to finish.
We would like to stop/end our script when the user selected to update git otherwise the git executable may not be available.
With returning 2 the calling program/script knows that the update is running and can do appropriate next steps.

Signed-off-by: Brodski, David <david.brodski@amd.com>