install.iss: check for trailing 'manager-core'
GCM Core uses a fully-scoped path for the 'credential.helper' config
key. If GCM Core set its value in the system config this way, then the
installer wizard shows the default credential helper as "None". This
then would overwrite the system config to not include any credential
helper, possibly breaking existing installations.

Change the default by also looking to see if the config matches at the
end of the string.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>