Revert "installer: allow configuring the default behavior of `git pull`"
This reverts commit 6ab14d9a169d7aeb3fdcf861148ec4d1a0de91cf.

The reason why `git pull` now aborts by default when a fast-forward is
not possible is to prevent accidental creation of merge commits when
rebasing would be more appropriate (or vice-versa). Encouraging users to
always set a different default globally (which they will do without
thinking) defeats the safeguard against such a mistake.

Signed-off-by: Alex Henrie <alexhenrie24@gmail.com>