git-credential-manager-core: rename it to git-credential-manager
The Git Credential Manager project was moved into its own GitHub org,
and used the opportunity to drop the "Core" from its name.

This would have conflicted with the previous package of the same name,
but that was packaging Git Credential Manager for Windows, which has
been superseded by what is now Git Credential Manager.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>