pr: fix detection of multiple touched package definitions
When running the PR builds, we first determine which packages' files
were touched, to know which ones need to be built.

However, that logic was broken when more than one package was touched,
due to the newline(s) in the value while trying to set the output
variable that will then give rise to the matrix build.

Let's fix this by not even having any newlines in the value.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>