installer: never treat inferred options as "unseen"
The installer looks at the system config to see whether the user has
changed any settings that the installer wants to configure. This
prevents choices made during the latest install overriding manual
configuration changes after said install.

Obviously, such an inferred setting was definitely seen by the user, and
we do not need to show them the respective installer page again.

However, due to a missing `begin ... end` pair, the code's indentation
(which reflected the intention) mismatched the code flow (which was
unintended).

Let's let the code flow as intended.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>