installer: offer to install Posh-Git
Posh-Git is a very neat PowerShell module that serves as a PowerShell
equivalent to Git's Bash prompt and tab completion.

Let's offer it as an optional component; When selected, it will be
installed from the PSGallery for all users and will then also be added
to the profile.

When selected, the uninstaller will also uninstall `posh-git`.

This addresses https://github.com/git-for-windows/git/issues/1384

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>