git-extra: un-hardcode even more Windows paths in /etc/xml/catalog
In 91a3a9715 (git-extra: un-hardcode Windows paths in /etc/xml/catalog,
2021-12-05), we already took care of the unwanted `C:/git-sdk-64-ci`
prefix of those paths.

However, we just switched from the Azure Pipeline that used to
synchronize the git-sdk-32/git-sdk-64 repositories with the Pacman
repositories to a pair of GitHub workflows to do the same, and the
latter use a different location, one that makes more sense for GitHub
Actions.

Let's catch that, too.

It is relatively easy, really: the drive prefix is a pretty good
tell-tale and all the paths we will want start with `/usr/` anyway. So
let's strip everything between those two anchors and we're good.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>