astextplain: pass non-word .doc files through again
Commit a45f9aba (textconv: Pass through non-MS-Word *.doc files, 2011-03-31) in
msysgit/msysgit introduced a way to pass files that antiword doesn't recognize
(likely plaintext files) to the caller of the textconv unmodified. When we imported
The file for Git for Windows version 2 we added a call to sed that broke this bypass.

Restore this feature and actually check for the antiword exit code again. As a bonus,
this allows us to only invoke sed if atiword succeeded.

Signed-off-by: Matthias Aßhauer <mha1993@live.de>