astextplain: pass through unrecognized .pdf/.do{cx,cm,tx,tm} files
We're signalling the caller a successful text conversion if we know the file extension,
so if our conversion command doesn't understand the file at hand we should at least
pass through the unmodified file to the caller. We already do this for some of the
known file extensions, so for a consistent UX we should have this fallback for all
files.

Since the output of pdftotext gets piped through sed, we use the same construct we use
with antiword to handle failures of pdftotext instead of sed.

Signed-off-by: Matthias Aßhauer <mha1993@live.de>