astextplain: supress pdftotext error messages
If we can convert the PDF to plain text, that's great. If we can't, we pass the file on
as-is. The error messages are really more warnings from our perspective, but since they
don't even inform the user about what file they're about, they're a nuisance in
operations with multiple text conversions.

This fixes https://github.com/git-for-windows/git/issues/3673

Signed-off-by: Matthias Aßhauer <mha1993@live.de>