ci: avoid a bogus error when no packages/artifacts need to be built
When a PR build detects that no artifact has to be built, not only is
there no job in the matrix, but the reader is also greeted with this
ugly message:

	Error when evaluating 'strategy' for job 'build-artifacts'.
	.github/workflows/main.yml (Line: 63, Col: 20): Matrix vector
	'directory' does not contain any values

See e.g.
https://github.com/git-for-windows/build-extra/actions/runs/1780156306

The same would happen if no packages need to be built.

Let's fix this by telling GitHub Actions to not even evaluate the matrix
if the respective list (artifacts, or packages) is empty.

Note: we need to add a bogus string because otherwise the parser that
wants to construct the matrix jobs will _still_ fail.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>