git-extra(vimrc): only default to UTF-8, don't enforce it
Due to a typo introduced already in the first revision of Git for
Windows' `/etc/vimrc` (see 2c38e4551 (Add two files that need to be
installed, 2015-02-25)), by mistake we enforce the commit message to be
written in UTF-8.

We did not want to override the valid file encodings, though, we just
wanted to set the file encoding to UTF-8 by default.

The trivial typo "fileencodings" (note the "s" at the end) caused the
former rather than the latter, though. For details, see
http://vimdoc.sourceforge.net/htmldoc/options.html#'fileencoding' and
http://vimdoc.sourceforge.net/htmldoc/options.html#'fileencodings'.

This fixes https://github.com/git-for-windows/git/issues/3725.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>