winpty: auto-detect redirects of stdin/stdout
Running commands through winpty with a redirected stdin or
stdout only leads to error messages, that are at best not very
helpfull to users and at worst outright confusing users.
We can do better by adding a thin wrapper script that detects
when it is and isn't appropriate to run a command through winpty
and acts accordingly.

Signed-off-by: Matthias Aßhauer <mha1993@live.de>