installer: allow building as administrator
With Git for Windows proper, this is not necessary, because there is
this work-around where a worktree owned by the Administrators group is
considered safe when running in elevated mode.

But we want to build a test installer based off of a tentative v2.30.5,
therefore we need a work-around in the script that is used to create the
Git for Windows installers.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>