make-file-list.sh: make it locale-independent
When running `grep --perl-regexp` with a non-ASCII or non-UTF-8 locale
(or, when no locale is specified at all!), it simply refuses to run.

Since we _know_ that we want to handle ASCII, or at most UTF-8, we can
force the locale for that `grep` invocation and ensure that it works all
of the time, even if run from a non-login shell (which is the case in
GitHub workflows).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>