make-file-list: do include p11-kit's dependencies
Apparently recent versions of the `p11-kit` package now depend on libffi
and on libtasn1, so let's not exclude them.

Unless we're generating the file list for MinGit, in which case p11-kit
is excluded.

Another complication is that the i686 flavor of `p11-kit` has not been
built in a long time, therefore we do _not_ have that dependency in
32-bit builds.

Pointed out by `check-for-missing-dlls.sh`.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>