ci: manually simulate the `setup-git-for-windows-sdk` Action
The problem with using the Action itself is that it clones `build-extra`
and uses `please.sh`/`make-file-list.sh` to generate the
`build-installers` flavor, but these two scripts might actually be
modified by a PR we might want to test.

In the concrete instance that motivated this commit, we are about to
include `libffi` and `libtasn` in `make-file-list.sh`, but running the
`setup-git-for-windows-sdk` Action would not include them just yet
(because `make-file-list.sh` is unmodified on the `main` branch but not
in the PR branch we want to test).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>