check-for-missing-dlls: make it an error if there are unused DLLs
Some "unused" DLLs will turn out not to be unused after all, e.g. when
they are loaded via `LoadLibraryW()` instead of being compile-linked.

However, we want to be intentional about those, just as we were with
`libstdbuf.dll`: we want to list those specifically in the
`check-for-missing-dlls` script as exceptions.

If the script reports anything unexpected, we really want to be notified
about this, which means that the script must fail so that the automated
build that runs it reports a failed run.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>