check-for-missing-dlls: only report progress when run interactively
It is a bit distracting to see all those directories in the output of a
GitHub workflow, and it makes it unnecessarily hard to find the relevant
bits in the output, i.e. the missing and unused DLLs.

So let's only report the directories as they are scanned when running
interactively (in which case we can overwrite the same line with the
directory names over and over again, which reduces the distraction
noticably).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>