check-for-missing-dlls: avoid `ldd` problem with `getprocaddr64.exe`
For technical reasons, `getprocaddr64.exe` is a x86_64 program, and when
the i686 flavor of `ldd.exe` is asked to look at this file, it has no
idea what to do and throws its hands in the air with this error message:

	ldd: /usr/libexec/getprocaddr64.exe: Function not implemented

This is not helpful.

The `getprocaddr64.exe` helper is compiled as a MINGW program with
minimal dependencies, though, therefore we can simply skip it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>