make-file-list: avoid warning about Pacman's sync databases
When building Git for Windows' installer & friends, we use a subset of
the Git for Windows SDK, specifically excluding files we do not actually
need. This includes the sync databases, i.e. the cached copy of the
Pacman repositories' files describing the current package versions.

Let's suppress the warning `pacman` displays (and has no way of turning
off, not even with the `--quiet` option) when those sync databases are
missing.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>