create-sdk-artifact build-installers: make sure that `/tmp/` exists
Otherwise `make-file-list.sh` will fail to run properly: we just
introduced a change to said script where it writes a temporary file.

And since `/etc/nsswitch.conf` is not (yet) installed, `/tmp/` is not
automatically mapped to `$TEMP`. But it is not necessary, either, as
we're not creating the SDK artifact in a read-only file location (in
contrast to the Git for Windows installer, which uses `C:\Program
Files\Git` as the pseudo Unix root).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>