make-file-list: avoid warning about Pacman's sync databases more consistently
In 5883116d6 (make-file-list: avoid warning about Pacman's sync
databases, 2022-07-09), a change was introduced that tries to avoid this
warning when Pacman's sync databases are not cached locally (such as is
the case in the `build-installers` SDK artifact).

However, the change was neither complete nor was it tested thoroughly:
the `/tmp/` directory may not exist yet when `make-file-list.sh` is run,
e.g. when generating the `build-installers` SDK itself.

Fix these issues properly.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>