please.sh build-mingw-w64-git: ensure small-ish source package
While publishing Git for Windows v2.37.3, the deployment to the Pacman
repository failed in the first attempt. The reason was that the source
package was larger than 256MB, which is not supported on Azure Blobs.

Let's reduce the size of the source package.

Granted, ~150MB is still not _quite_ small, but it leaves us quite a bit
more wiggle room with respect to Azure Blobs' limitations. And it's the
best we can do for now: The source package contains the complete
git-for-windows/git repository, after all.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>