please.sh upgrade busybox: allow MINGW `git` to access the source tree
Some time through the past years, `makepkg` was changed so that it uses
the bare clone of the Git repository as an alternate for the actual
checkout, but it uses `/usr/src/git` to install that alternate, which
means that it uses a Unix-style absolute path.

That don't work with MINGW Git, which requires Windows-style paths.

Let's side-step the problem and dissociate the worktree, like we do when
building the MSYS2 runtime.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>