pacman-helper quick_add: be smarter about determining a package's architecture
The `pacman-helper.sh` script can be used to deploy packages to Git for
Windows' Pacman repository. To do that, it needs to sort the packages
into the correct subdirectory, based on the package's architecture.

Previously, when the filename did not reveal for which architecture the
package was built, we used the environment variable
`FALLBACK_ARCHITECTURE` to specify it.

This won't work going forward because we want to deploy the same package
for multiple architectures in a single call, for which a single fallback
just won't cut it.

Let's use the information in the `.BUILDINFO` file that is contained in
every Pacman package to derive the architecture even in cases where the
file name does not reveal it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>