please create-sdk-artifact: add a flavor focused on Pacman
With the current push to automate All The Things in Git for Windows, we
need a way to perform tasks that involve Pacman. It would be a bit
wasteful to use a full Git for Windows SDK for that task, so here is
support for an SDK artifact that includes Pacman itself and its
dependencies, but leaves out other packages (such as `make`, `gcc`,
etc).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>