please create-sdk-artifact: slim down the `pacman` artifact
This trims the output to (currently) ~100MB, by skipping files that are
not strictly necessary to run `pacman`, such as documentation or header
files.

On the off chance that those files _are_ needed in some scenarios, offer
also the `pacman-full` flavor, which does include all the files that are
promised by the packages' file list (think: `pacman -Ql <pkg>`).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>