installer/portable/mingit: stop shipping `libssp-0.dll` unless needed
A recent GCC update made that library obsolete (it used to be required
to make `-fstack-protector` work). However, we had not re-built
`git.exe` with that newer GCC version, therefore we still needed to ship
it.

Now that Git v2.39.0-rc0 is available, we have a new `git.exe` that no
longer requires `libssp-0.dll`, and since it was the last user of that
library, now the `check-for-missing-dlls` check fails, complaining that
the library should no longer be included.

We have to be careful to keep the script working even when a `git.exe`
is shipped that is built using an older GCC: every once in a blue moon,
the Git for Windows project needs to release MinGit backports, and those
are traditionally done using the very snapshot of the Git for Windows
SDK with which that release train was started. And that snapshot will
still have an older GCC.

To accommodate for that, we do not simply exclude the `.dll` but ensure
that the `git.exe` does not need it. If it needs it, we include it.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>