add-release-note: remove superseded entries
In 8f139e6c7 (please.sh: script to fix duplicate relnote entry,
2018-11-06), in what must be described as an heroic effort to read a
`sed` script that was crafted by yours truly, to understand it and to
modify it correctly, the original `please.sh mention [...]` command was
taught to remove superseded entries while adding new entries to the "New
Features" section.

In Javascript, this is naturally much easier to write, to understand and
to modify.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>