ci: add a workflow to help automating the release notes updates
With this workflow, we can now hand off the nitty-gritty of adding a
release note to GitHub Actions. This opens the door for more automation
in the future, hopefully eventually being as simple as adding an issue
comment `/relnote <type> <message>` to any of Git for Windows' tickets.

For the moment, though, this is still a manual process where the user
has to trigger a `workflow_dispatch` with the relevant information
provided in the form of inputs.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>